Refactor authentication handlers and user handlers
- Refactor authentication handlers to use updated response schemas for authentication responses.
- Remove unused dependencies and import statements.
- Refactor user handlers to include endpoints for fetching user details and uploading profile images.
- Update OpenAPI documentation to include a new "users" section.
